requirements.txt
crewai==1.14.1
crewai-tools==1.14.1
pydantic>=2.8,<3.0
PyYAML>=6.0,<7.0
python-dotenv>=1.0,<2.0
pandas>=2.2,<3.0
rich>=13.7,<14.0
pytest>=8.0,<9.0
